Update update_app.sh
With the previous version, after I executed the update script, and logged out of the AWS EC2 machine, the DoccoLink repository would get deleted. The cause for this were the `read` instructions that had the `-s` flag. The `-s` flag does not show the password while the user types it, so I assume there was some problem with `sed` while trying to append the hidden inputs to the `web_variables.env` file. With this version of the script, the passwords will not be hidden, but the DoccoLink repository will not get deleted after logging out.